Fix wrongly decoded URL in Safari
We received bugreports from some people that they can't checkin.
After closer examination, there are two problems coming together:

   - some QR Code Scanners encode the URL params before redirecting to
     the browser
   - iOS Safari decodes already decoded params wrong with
     URLSearchParams

This means we can't use URLSearchParams to parse the url params. Thus
I replaced it with a regex. Yes, a regex. It's not optimal because it
requires that the order of the params is fixed (?a=...&k=...), but
it fixes the problem.